+++ django_prose_editor/widgets.py
@@ -3,7 +3,12 @@ from django import forms
 
 class ProseEditorWidget(forms.Textarea):
     class Media:
-        css = {"screen": ["django_prose_editor/editor.css"]}
+        css = {
+            "screen": [
+                "django_prose_editor/material-icons.css",
+                "django_prose_editor/editor.css",
+            ]
+        }
         js = [
             "django_prose_editor/editor.js",
             "admin/js/jquery.init.js",